BAEL-16781 Moving top articles to their respective packages (#7594)
